triangulation/requirements.txt
asyncpg==0.22.0
confluent-kafka==1.6.0
orjson==3.4.8
python-decouple==3.4